+++ lib/danger/commands/pr.rb
@@ -63,6 +63,7 @@ module Danger
           Danger::EnvironmentManager.danger_head_branch,
           @dangerfile_path,
           nil,
+          nil,
           nil
         )
       end