+++ lib/danger/commands/runner.rb
@@ -38,6 +38,7 @@ module Danger
       @head = argv.option("head")
       @fail_on_errors = argv.option("fail-on-errors", false)
       @new_comment = argv.flag?("new-comment")
+      @remove_previous_comments = argv.flag?("remove-previous-comments")
       @danger_id = argv.option("danger_id", "danger")
       @cork = Cork::Board.new(silent: argv.option("silent", false),
                               verbose: argv.option("verbose", false))
@@ -58,7 +59,8 @@ module Danger
         ["--fail-on-errors=<true|false>", "Should always fail the build process, defaults to false"],
         ["--dangerfile=<path/to/dangerfile>", "The location of your Dangerfile"],
         ["--danger_id=<id>", "The identifier of this Danger instance"],
-        ["--new-comment", "Makes Danger post a new comment instead of editing its previous one"]
+        ["--new-comment", "Makes Danger post a new comment instead of editing its previous one"],
+        ["--remove-previous-comments", "Removes all previous comment and create a new one in the end of the list"]
       ].concat(super)
     end
 
@@ -69,7 +71,8 @@ module Danger
         dangerfile_path: @dangerfile_path,
         danger_id: @danger_id,
         new_comment: @new_comment,
-        fail_on_errors: @fail_on_errors
+        fail_on_errors: @fail_on_errors,
+        remove_previous_comments: @remove_previous_comments
       )
     end
   end